refactor: 체크된 상품이 없으면 dispatch를 실행하지 않도록 수정

+++ src/pages/cart/components/product-list/ProductList.jsx
@@ -23,6 +23,11 @@ function ProductList() {
   };
 
   const handleClick = () => {
+    const isNonExistCheckedItem = cartList.every(
+      (item) => item.checked === false
+    );
+    if (isNonExistCheckedItem) return;
+
     dispatch(removeCheckedCartItem());
   };
